Add script and icon
Since I used pyinstaller to package the script, the executable file is 40MB sized, which exceeds the limit of an uploadable file on GitHub. I'm still figuring out a way to upload a file that big.  For now, you can package your own executable file or just run the script directly.